+++ apps/messages-app/CHANGELOG.md
@@ -1,5 +1,17 @@
 # messages-app
 
+## 0.2.4
+
+### Patch Changes
+
+- c5e169e: Add logs through @pagopa/azure-tracing
+
+## 0.2.3
+
+### Patch Changes
+
+- 23b00ca: Remove at alias for src directory
+
 ## 0.2.2
 
 ### Patch Changes